Fixes issue where guard-sass would be fired after css syntax errors
Upon saving a file with syntax errors, this section of code was
failing, and causing guard-sass to be fired. Specifically, a
"NoMethodError: undefined method `intersection' for []:Array"
would be triggered.

The fix: If an exception is raised while trying to get the
dependencies of a file, move on to the next file. And do not
try to find the intersection from the dependencies that we
failed to get.